refactor: checkbox into single component
