+++ packages/ui/src/lib/checkBox/Introduction.mdx
@@ -1,0 +1,9 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Ui/Checkbox" />
+
+# Checkbox
+
+This provides a (square) checkbox component that can be colored to act as both a categorical color legend, and a control for which categories of things are displayed on a map or visualization.
+
+It should be used in cases where **multiple** items from a list of options can be selected simultaneously; if only **one** item can be selected at a time, then a round [RadioButton](./?path=/docs/ui-radiobutton--docs) should be used instead.
