--- packages/ui/src/lib/coloredCheckbox/Introduction.mdx
@@ -1,9 +1,0 @@
-import { Meta } from '@storybook/blocks';
-
-<Meta title="Ui/ColoredCheckbox" />
-
-# Colored checkbox
-
-This provides a (square) colored checkbox component that can act as both a categorical color legend, and a control for which categories of things are displayed on a map or visualization.
-
-It should be used in cases where **multiple** items from a list of options can be selected simultaneously; if only **one** item can be selected at a time, then a round [ColoredRadioButton](./?path=/docs/ui-coloredradiobutton--docs) should be used instead.
